Divided into modules
разбила код на модули и создала файл api.js, в котором находятся все запросы
